refactor: colocate types with their owning features
