+++ src/index.ts
@@ -1,7 +1,35 @@
 export * from "./api";
 export * as Columns from "./columns";
-export { Progress } from "./services/progress";
-export * from "./services/stdio";
-export * from "./types";
+export { Progress, type ProgressService } from "./services/progress";
+export { ProgressStdio, type ProgressStdioService } from "./services/stdio";
 export { useNow } from "./renderer/context/now-context";
 export { useSpinnerTick } from "./renderer/context/spinner-context";
+export {
+  TaskId,
+  TaskStatusSchema,
+  type TaskStatus,
+  TaskCountDisplaySchema,
+  type TaskCountDisplay,
+  TaskUnitsSchema,
+  type TaskUnits,
+  TaskProgressSampleSchema,
+  type TaskProgressSample,
+  TaskSnapshotSchema,
+  type TaskSnapshot,
+} from "./task-model";
+export {
+  type ColumnAlign,
+  type TaskTreeInfo,
+  type TaskRowDerived,
+  type CellInfo,
+  type ColumnRenderContext,
+  type ColumnSizeValue,
+  type ColumnDef,
+  type Column,
+} from "./columns/types";
+export { type TaskHandle } from "./tasks/task-handle";
+export { type AddTaskOptions, type UpdateTaskOptions, type TrackOptions } from "./tasks/options";
+export { type RenderRow, type TaskStore } from "./services/store/types";
+export { type TaskOperations } from "./services/task-operations";
+export { Task } from "./tasks/current-task";
+export { type TaskApi } from "./tasks/task-api";
