+++ src/tasks/current-task.ts
@@ -1,0 +1,6 @@
+import { Context } from "effect";
+import type { TaskId } from "../task-model";
+
+export class Task extends Context.Service<Task, TaskId>()(
+  "stromseng.dev/effective-progress/Task",
+) {}
